+++ src/Kernel/Support/Helpers.php
@@ -1,0 +1,128 @@
+<?php
+
+/*
+ * This file is part of the Easeava package.
+ *
+ * (c) Easeava <tthd@163.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace EaseBaidu\Kernel\Support;
+
+use Illuminate\Support\Str;
+
+/**
+ * Generate a signature.
+ *
+ * @param array  $attributes
+ * @param string $key
+ * @param string $encryptMethod
+ *
+ * @return string
+ */
+function generate_sign(array $attributes, $key, $encryptMethod = 'md5')
+{
+    ksort($attributes);
+
+    $attributes['key'] = $key;
+
+    return strtoupper(call_user_func_array($encryptMethod, [urldecode(http_build_query($attributes))]));
+}
+
+/**
+ * Generate a signature.
+ *
+ * @param array  $attributes
+ * @param string $rsaPriKeyStr
+ *
+ * @return string
+ */
+function generate_sign_with_rsa(array $attributes, $rsaPriKeyStr)
+{
+    $priKey = openssl_pkey_get_private($rsaPriKeyStr);
+    ksort($attributes);
+    openssl_sign(urldecode(http_build_query($attributes)), $sign, $priKey);
+    openssl_free_key($priKey);
+
+    return base64_encode($sign);
+}
+
+/**
+ * Get client ip.
+ *
+ * @return string
+ */
+function get_client_ip()
+{
+    if (! empty($_SERVER['REMOTE_ADDR'])) {
+        $ip = $_SERVER['REMOTE_ADDR'];
+    } else {
+        // for php-cli(phpunit etc.)
+        $ip = defined('PHPUNIT_RUNNING') ? '127.0.0.1' : gethostbyname(gethostname());
+    }
+
+    return filter_var($ip, FILTER_VALIDATE_IP) ?: '127.0.0.1';
+}
+
+/**
+ * Get current server ip.
+ *
+ * @return string
+ */
+function get_server_ip()
+{
+    if (! empty($_SERVER['SERVER_ADDR'])) {
+        $ip = $_SERVER['SERVER_ADDR'];
+    } elseif (! empty($_SERVER['SERVER_NAME'])) {
+        $ip = gethostbyname($_SERVER['SERVER_NAME']);
+    } else {
+        // for php-cli(phpunit etc.)
+        $ip = defined('PHPUNIT_RUNNING') ? '127.0.0.1' : gethostbyname(gethostname());
+    }
+
+    return filter_var($ip, FILTER_VALIDATE_IP) ?: '127.0.0.1';
+}
+
+/**
+ * Return current url.
+ *
+ * @return string
+ */
+function current_url()
+{
+    $protocol = 'http://';
+
+    if ((! empty($_SERVER['HTTPS']) && 'off' !== $_SERVER['HTTPS']) || ($_SERVER['HTTP_X_FORWARDED_PROTO'] ?? 'http') === 'https') {
+        $protocol = 'https://';
+    }
+
+    return $protocol.$_SERVER['HTTP_HOST'].$_SERVER['REQUEST_URI'];
+}
+
+/**
+ * Return random string.
+ *
+ * @param string $length
+ *
+ * @return string
+ */
+function str_random($length)
+{
+    return Str::random($length);
+}
+
+/**
+ * @param string $content
+ * @param string $publicKey
+ *
+ * @return string
+ */
+function rsa_public_encrypt($content, $publicKey)
+{
+    $encrypted = '';
+    openssl_public_encrypt($content, $encrypted, openssl_pkey_get_public($publicKey), OPENSSL_PKCS1_OAEP_PADDING);
+
+    return base64_encode($encrypted);
+}
