+++ src/Service/Payment/Application.php
@@ -3,7 +3,6 @@
 /**
  * EaseBaidu - A PHP Framework For Baidu SDK
  *
- * @package  Easeava/baidu
  * @author   Easeava <tthd@163.com>
  */
 
@@ -11,9 +10,14 @@ namespace EaseBaidu\Service\Payment;
 
 use EaseBaidu\Kernel\Container;
 
+/**
+ * Payment Application.
+ *
+ * @property \EaseBaidu\Service\Payment\Jssdk\Client             $jssdk
+ */
 class Application extends Container
 {
     protected $providers = [
-
+        Jssdk\ServiceProvider::class,
     ];
 }
